+++ internal/api/admin/func_listadminmenu.go
@@ -49,7 +49,7 @@ func (h *handler) ListAdminMenu(ctx *gin.Context) {
 	}
 
 	searchOneData := new(admin.SearchOneData)
-	searchOneData.Id = int32(ids[0])
+	searchOneData.Id = int64(ids[0])
 	searchOneData.IsUsed = 1
 
 	info, err := h.adminService.Detail(ctx, searchOneData)
@@ -62,7 +62,7 @@ func (h *handler) ListAdminMenu(ctx *gin.Context) {
 	res.UserName = info.Username
 
 	searchData := new(admin.SearchListMenuData)
-	searchData.AdminId = int32(ids[0])
+	searchData.AdminId = int64(ids[0])
 
 	listData, err := h.adminService.ListMenu(ctx, searchData)
 	if err != nil {
