+++ internal/api/authorized/func_create.go
@@ -19,7 +19,7 @@ type createRequest struct {
 }
 
 type createResponse struct {
-	Id int32 `json:"id"` // 主键ID
+	Id int64 `json:"id"` // 主键ID
 }
 
 // Create 新增调用方
