+++ internal/api/cron/func_create.go
@@ -30,7 +30,7 @@ type createRequest struct {
 }
 
 type createResponse struct {
-	Id int32 `json:"id"` // 主键ID
+	Id int64 `json:"id"` // 主键ID
 }
 
 // Create 创建任务
