+++ internal/services/admin/service_create.go
@@ -16,7 +16,7 @@ type CreateAdminData struct {
 	Password string // 密码
 }
 
-func (s *service) Create(ctx context.Context, adminData *CreateAdminData) (id int32, err error) {
+func (s *service) Create(ctx context.Context, adminData *CreateAdminData) (id int64, err error) {
 	model := admin.NewModel()
 	model.Username = adminData.Username
 	model.Password = password.GeneratePassword(adminData.Password)
