+++ internal/services/cron/service_execute.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ChangSZ/mall-go/internal/repository/mysql/cron_task"
 )
 
-func (s *service) Execute(ctx context.Context, id int32) (err error) {
+func (s *service) Execute(ctx context.Context, id int64) (err error) {
 	qb := cron_task.NewQueryBuilder()
 	qb.WhereId(mysql.EqualPredicate, id)
 	info, err := qb.QueryOne(mysql.DB().GetDbR().WithContext(ctx))
