+++ internal/services/cron/service_updateused.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/cast"
 )
 
-func (s *service) UpdateUsed(ctx context.Context, id int32, used int32) (err error) {
+func (s *service) UpdateUsed(ctx context.Context, id int64, used int32) (err error) {
 	data := map[string]interface{}{
 		"is_used":      used,
 		"updated_user": core.SessionUserInfo(ctx).UserName,
