+++ internal/services/menu/service_delete.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ChangSZ/mall-go/internal/repository/mysql/menu"
 )
 
-func (s *service) Delete(ctx context.Context, id int32) (err error) {
+func (s *service) Delete(ctx context.Context, id int64) (err error) {
 	data := map[string]interface{}{
 		"is_deleted":   1,
 		"updated_user": core.SessionUserInfo(ctx).UserName,
