+++ internal/services/menu/service_updatesort.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ChangSZ/mall-go/internal/repository/mysql/menu"
 )
 
-func (s *service) UpdateSort(ctx context.Context, id int32, sort int32) (err error) {
+func (s *service) UpdateSort(ctx context.Context, id int64, sort int32) (err error) {
 	data := map[string]interface{}{
 		"sort":         sort,
 		"updated_user": core.SessionUserInfo(ctx).UserName,
